Created new classes
Created new classes(Database, DataTotal, GroupDates)
