+++ src/main/Summary.java
@@ -1,6 +1,8 @@
 package main;
 
 import processeddata.DataGroup;
+import processeddata.DataTotal;
+import processeddata.GroupDates;
 import util.ExceptionManager;
 import util.UserInputManager;
 
@@ -61,25 +63,25 @@ public class Summary {
         int groupingCondition = UserInputManager.getIntUserInput();
 
         LocalDate[] userTimeRange = userDataObj.getTimeRange();
-        DataGroup baseDayGroup = Data.setDateList(userTimeRange);
+        DataGroup baseDayGroup = GroupDates.setDateList(userTimeRange);
         ArrayList<DataGroup> groupedDayList = new ArrayList<>();
 
         switch (groupingCondition) {
-            case 1 -> groupedDayList = Data.noGrouping(baseDayGroup);
+            case 1 -> groupedDayList = GroupDates.noGrouping(baseDayGroup);
             case 2 -> {
                 UserInputManager.displayMenu("Please enter the number of groups you want to create. (Integer value)>> ");
                 int numOfGroups = UserInputManager.getIntUserInput();
-                groupedDayList = Data.groupByGroupNum(baseDayGroup, numOfGroups);
+                groupedDayList = GroupDates.groupByGroupNum(baseDayGroup, numOfGroups);
             }
             case 3 -> {
                 UserInputManager.displayMenu("Please enter the number of days in a group. (Integer value)>> ");
                 int numOfDays = UserInputManager.getIntUserInput();
-                groupedDayList = Data.groupByDayNum(baseDayGroup, numOfDays);
+                groupedDayList = GroupDates.groupByDayNum(baseDayGroup, numOfDays);
             }
             default -> ExceptionManager.throwInvalidOption();
         }
 
-        ArrayList<DataGroup> analyzedData = processeddata.Data.getData(userDataObj, groupedDayList);
+        ArrayList<DataGroup> analyzedData = processeddata.Database.getData(userDataObj, groupedDayList);
 
         // FIXME: 2021-08-10 Lee Gain
         System.out.println("Analyzed result+++++++++++++++++++++++++++++++++++++++++++");
@@ -122,8 +124,8 @@ public class Summary {
             int value = 0;
 
             switch (resultType) {
-                case 1 -> value = getNewTotal(dtArr, metric);
-                case 2 -> value = getUpTo(dtArr, metric);
+                case 1 -> value = DataTotal.getNewTotal(dtArr, metric);
+                case 2 -> value = DataTotal.getUpTo(dtArr, metric);
                 default -> ExceptionManager.throwInvalidOption();
             }
 
@@ -149,42 +151,5 @@ public class Summary {
             System.out.println(s);
         }
     }
-
-    private static int getNewTotal(ArrayList<processeddata.Data> db, int metric) {
-        // This method accepts array list of Data in processeddata package "db" and integer "metric"
-        // and returns a total of new cases of the data based on the metric that the user inputs.
-        // It throws an exception if the user input is invalid.
-
-        int value = 0;
-        for(processeddata.Data dt : db) {
-            switch (metric) {
-                case 1 -> value += dt.getNewCases();
-                case 2 -> value += dt.getNewDeaths();
-                case 3 -> value += dt.getNewPeopleVaccinated();
-                default -> ExceptionManager.throwInvalidOption();
-            }
-        }
-        return value;
-    }
-
-    private static int getUpTo(ArrayList<processeddata.Data> db, int metric) {
-        // This method accepts array list of Data in processeddata package "db" and integer "metric"
-        // and returns accumulated total data based on the metric that the user inputs.
-        // It throws an exception if the user input is invalid.
-
-        int value = 0;
-        int dbLength = db.toArray().length;
-        processeddata.Data lastDateData = db.get(dbLength - 1);
-
-        switch (metric) {
-            case 1 -> value = lastDateData.getTotalCases();
-            case 2 -> value = lastDateData.getTotalDeaths();
-            case 3 -> value = lastDateData.getPeopleVaccinated();
-            default -> ExceptionManager.throwInvalidOption();
-        }
-
-        return value;
-    }
-
 
 }
